refactor(app): clean up code and improve idiomatic Rust usage
Refactor various modules to improve code quality, simplify logic, and
adhere to Rust best practices:
- Implement `Default` for `AppStateTracker`, `OcrDetector`, and
  `SyncSortMode` using derive macros.
- Simplify conditional logic and control flow using `?` operator,
  `contains`, and combined boolean expressions.
- Optimize string and character handling in CLI tools and UI reporting.
- Clean up redundant type casting and improve variable access patterns.
- Minor UI/UX logic refinements in `overlay_ui` and `native_app_viewports`.

+++ rust/overmax_app/src/ui/settings_ui.rs
@@ -143,7 +143,7 @@ fn overlay_section(ui: &mut egui::Ui, draft: &mut Value) {
         let (rect, _response) = ui.allocate_exact_size(egui::vec2(280.0, 120.0), egui::Sense::hover());
         ui.painter().rect(
             rect,
-            egui::CornerRadius::same(Theme::R_SM as u8),
+            egui::CornerRadius::same(Theme::R_SM),
             egui::Color32::from_black_alpha(220),
             egui::Stroke::new(1.0, Theme::STROKE),
             egui::StrokeKind::Inside,
